Remove more progression code
